requirements.txt
python-dotenv==1.0.0
openai==1.12.0
PyPDF2==3.0.1
python-json-logger==2.0.7
email-validator==2.0.0.post2
redis==5.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
fastapi==0.109.1
uvicorn==0.27.0
pytest==7.4.4
spacy>=3.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz